refactor(forms): merge runValidation and validate
Move runValidation from validateField to validate.

As runValidation becomes the default export

rename file to runValidation.js.

Reuse runValidation in validateForm and configureField.
