+++ packages/cerebral-forms/src/utils/configureField.js
@@ -1,5 +1,4 @@
-import checkHasValue from './checkHasValue'
-import validate from './validate'
+import runValidation from './runValidation'
 
 export default function configureField (formData, field) {
   // If not an actual field but a global property or just a namespace
@@ -14,25 +13,22 @@ export default function configureField (formData, field) {
   const validationRules = field.validationRules || null
   const validationMessages = field.validationMessages || []
   const requiredMessage = field.requiredMessage || null
-  const hasValue = checkHasValue(formData, value, isValueRules)
-  const validationResult = validate(formData, value, validationRules)
 
   field.defaultValue = defaultValue
   field.validationRules = validationRules
   // Field is valid only when there is a value and the validation rule
   // says it is valid. If "isRequired" it will only be valid if it actually
   // has a value
-  field.isValid = ((hasValue && validationResult.isValid) || (!isRequired && !hasValue))
   field.validationMessages = validationMessages
   field.requiredMessage = requiredMessage
-  if (isRequired && !hasValue) {
-    field.errorMessage = requiredMessage
-  } else {
-    field.errorMessage = validationResult.isValid ? null : validationMessages[validationResult.failedRuleIndex]
-  }
   field.isValueRules = isValueRules
   field.isRequired = isRequired
-  field.hasValue = hasValue
+
+  const validationResult = runValidation(field, formData)
+
+  field.isValid = validationResult.isValid
+  field.errorMessage = validationResult.errorMessage
+  field.hasValue = validationResult.hasValue
   field.isPristine = true
 
   return field
