+++ packages/cerebral-forms/src/utils/runValidation.js
@@ -1,6 +1,7 @@
+import checkHasValue from './checkHasValue'
 import rules from '../rules.js'
 
-export default function validate (form, value, validations) {
+function validate (form, value, validations) {
   const initialValidation = {
     isValid: true,
     failedRuleIndex: null
@@ -47,5 +48,26 @@ export default function validate (form, value, validations) {
         failedRuleIndex: index
       }
     }, initialValidation)
+  }
+}
+
+export default function runValidation (field, form) {
+  const hasValue = checkHasValue(form, field.value, field.isValueRules)
+  const result = validate(form, field.value, field.validationRules)
+  const isValid = result.isValid && (
+    (field.isRequired && hasValue) ||
+    !field.isRequired
+  )
+  const errorMessage = field.isRequired && !hasValue
+    ? field.requiredMessage
+    : result.isValid
+        ? null
+        : field.validationMessages[result.failedRuleIndex]
+
+  return {
+    isValid,
+    isPristine: false,
+    hasValue: checkHasValue(form, field.value, field.isValueRules),
+    errorMessage: errorMessage
   }
 }
